beacon-package/requirements.txt
python-dotenv~=1.0.0
azure-cognitiveservices-speech
selenium
SpeechRecognition
pyaudio~=0.2.11
g4f
